Transpile try words at level 1
The try word calls the function on the top of the stack. If it raises an exception, it looks within the handlers that were initialially below the function in the stack. The handlers are an iterable of pairs. The first of each pair is a class or tuple of classes. The second of each pair is a function. Handler selection is similar to that of Python. If there is no suitable handler, the exception is re-raised.