Don't describe the top-level parser with the arg to parsy.generate
That causes finer-grained error info to be thrown away. Instead, we now
*add* the description to the original parsing failure. The error
messages aren't great, but it's an improvement that's an aid in
debugging.

To do this, we create a parser combinator desc_cumulatively that adds
its 'expected' argument to the Result object returned by the original
parser instead of replacing the Result object (and getting rid of the
error position). We also add to the parsy type stub for type checking
purposes.