Add location info to all type checker errors (level 1)
All errors that can be raised during the normal course of type checking
(type errors and name errors) now have the location in the Concat source
code of the occurrence attached to them. The location is diaplayed,
along with the line of the error, in a user friendlier error message
when running a script from a file.