Write NameWordNode case in more 'bidirectional' style (level 1)
To be honest, it isn't really looking bidirectional...

Anyway, this commit introduces a more advanced constraint system that
can check subtyping constraints. It also introduces an abstraction
TypeSequence for sequences of types that is a subclass of Type.
Using TypeSequence is far more convenient than implementing functions
on top of Python's built-in sequence types.