Remove the _IntersectionType case from unify (level 1)
I currently think that having intersection types is not worth the cost of
the extra complexity. Thus, code related to it has been removed from the
unifier, and a TODO has been added to decide whether to remove intersection
types fully. They haven't been removed yet because a lot of code uses them.