Get rid of the unifier entirely (levels 1 & 2)
Now, we use a constraint solving system where the constraints are subtyping
relationships. We also generally don't use any sort of constraint to
propagate type information backwards except for subtyping, so if, say, you
try to access an attribute of an object on the stack, the type checker must
already know that there is a object with that attribute at the top of the
stack. It won't try to infer that information by constraining a sequence
variable currently on the stack.

# Conflicts:
#	concat/level1/typecheck/__init__.py