Improve the speed of the type checker on tests (level 1 and 2)
Changes recent in the commit history slowed down the type checker to the
point that waiting for tests to finish was unreasonable. On my machine,
the tests took about 3000 seconds to finish, even with some of the
slowest of the tests disabled. The tests now take 30-40 seconds (with
the same tests disabled) because of the following modifications:

* Removing the special sequence case from Substitutions.__call__

It seems that checking if an object is an instance of an abstract class
(collections.abc.Sequence, in this case) can take quite a while over
many iterations.

* Not substituting the head of an object type when substituting the
  object type

Doing otherwise results in long head chains that get traversed every
substitution.

These changes required fixes to various tests, including
`test_add_operator_inference` and
`concat.tests.level1.stdlib.test_repl.TestREPLFunctions.test_read_quot`.

Additionally, the _Function class now uses TypeSequences instead of
plain sequences to catch errors where a sequence variable appears in a
sequence type outside of the first position earlier.