Compare nominal types for equality by identity (level 1)
To make sure that a nominal type doesn't become duplicated when it
shouldn't, substitution doesn't create a new type object when no free type
variables are in the domain of the substitution (i.e. when the substitution
won't change the type anyway). Also, instiantations of an ObjectType are
not recomputed.

I think this change halfed the time it takes to type check the 'symmetric
tree' example.