Replace all uses of `_global_constraints.add` with algorithmic subtyping (levels 1 & 2)
I think encountered the term 'algorithmic subtyping' in the
r/programminglanguages Discord, and found that's described in Types and
Programming Languages. This is kind of a return to my previous approach to
subtyping. My constraint solver is currently the speed bottleneck, and so
far using algorithmic subtyping instead has improved the speed of the tests
by a factor of 4, on my machine.

I also inadvertently fixed some bugs related to pushed quotations and
subscriptions, and edited the tests accordingly.