Define type equality using antisymmetry of subtype relation (levels 1 & 2)
In the course of this patch, I found a bug in the way stack effect type
nodes are converted to types that resulted in a type like (*i -- *i)
containing two distinct variables. Subtyping of type constructors is
implemented, but I need to create an ordered set implementation to help
me with renaming bound variables.