Tokenize all operators as names
To get the tests passing again, I had also modify the type checker to
understand that the operators exist in the built-in environment. Also, I
improved the type definitions for iterables are iterators, and I staved off
infinite recursion in the subtyping algorithm using the idea of 'traces'
from "Subtyping Recursive Types" (Amadio & Cardelli, 1993).